add RTL support and system UI styling, update app theme and constants

+++ lib/core/constants/app_constants.dart
@@ -1,6 +1,6 @@
 class AppConstants {
-  static const int appWidth = 390;
-  static const int appHeight = 844;
+  static const double appWidth = 390;
+  static const double appHeight = 844;
 
   static const String hasSeenOnboardingKey = 'has_seen_onboarding';
